new file:   build/_calculix_pythoninterface.so 	renamed:    src/Core/CalculiXPythonInterface.py -> build/calculix_pythoninterface.py 	modified:   build/libcalculix_plugin.ccm 	modified:   src/CMakeLists.txt 	modified:   src/Core/CMakeLists.txt 	deleted:    src/Core/CalculiXPythonInterface.cpp 	deleted:    src/Core/CalculiXPythonInterface.i 	new file:   src/PythonInterface/CMakeLists.txt 	new file:   src/PythonInterface/CalculiXPythonInterface.cpp 	renamed:    src/Core/CalculiXPythonInterface.hpp -> src/PythonInterface/CalculiXPythonInterface.hpp 	new file:   src/PythonInterface/CalculiXPythonInterface.i 	renamed:    src/Core/CalculiXPythonInterface_wrap.cxx -> src/PythonInterface/CalculiXPythonInterface_wrap.cxx 	new file:   src/PythonInterface/calculix_pythoninterface.py
