--- src/Core/CalculiXPythonInterface.cpp
@@ -1,18 +1,0 @@
-#include "CalculiXPythonInterface.hpp"
-//#include "CalculiXCoreInterface.hpp"
-
-CalculiXPythonInterface::CalculiXPythonInterface()
-{
-  //CalculiXCoreInterface *ccx_iface = new CalculiXCoreInterface();
-}
-
-CalculiXPythonInterface::~CalculiXPythonInterface()
-{
-  //delete ccx_iface;
-}
-
-std::string  CalculiXPythonInterface::to_string_scientific(double value, int precision)
-{
-  //return ccx_iface->to_string_scientific(value, precision);
-  return "python test";
-}
