--- src/Core/CalculiXPythonInterface.i
@@ -1,6 +1,0 @@
-%module CalculiXPythonInterface
-%{
-#include "CalculiXPythonInterface.hpp"
-%}
-
-%include "CalculiXPythonInterface.hpp"
