+++ src/PythonInterface/CalculiXPythonInterface.i
@@ -1,0 +1,5 @@
+%module calculix_pythoninterface
+%{
+    #include "CalculiXPythonInterface.hpp"
+%}
+%include "CalculiXPythonInterface.hpp"
